Refactor scores export tests into helpers and focused modules.
Extract shared scheduler/persistence factories, split materialize vs ensure
coverage, and test cached admission via public payload resolver APIs.
